requirements.txt
Click==8.1.3
python-dotenv==0.20.0
setuptools
